Extract categories
Extract XxCategory classes from Category class.
