+++ LifeCategory.java
@@ -1,0 +1,7 @@
+package JavaReport;
+
+public class LifeCategory {
+    public LifeCategory() {
+        System.out.println("1.철학");
+    }
+}
